refactor: adjust summary step component
